introduce TransactionDetails
