+++ app/components/TransactionElement/index.js
@@ -4,15 +4,15 @@ import { Clipboard, TouchableOpacity, StyleSheet, Text, View, Image } from 'reac
 import { colors, fontStyles } from '../../styles/common';
 import { strings } from '../../../locales/i18n';
 import { toLocaleDateTime } from '../../util/date';
-import { renderFromWei, weiToFiat, hexToBN, isBN, toBN, toGwei } from '../../util/number';
+import { renderFromWei, weiToFiat, hexToBN } from '../../util/number';
 import { toChecksumAddress } from 'ethereumjs-util';
 import Identicon from '../Identicon';
 import { connect } from 'react-redux';
 import Icon from 'react-native-vector-icons/FontAwesome';
 import { getActionKey } from '../../util/transactions';
-import { renderFullAddress } from '../../util/address';
 import { getNetworkTypeById } from '../../util/networks';
 import { getEtherscanTransactionUrl } from '../../util/etherscan';
+import TransactionDetails from './TransactionDetails';
 
 const styles = StyleSheet.create({
 	row: {
@@ -83,13 +83,6 @@ const styles = StyleSheet.create({
 		width: 24,
 		height: 24
 	},
-
-	detailRowWrapper: {
-		flex: 1,
-		backgroundColor: colors.concrete,
-		paddingVertical: 10,
-		paddingHorizontal: 15
-	},
 	detailRowTitle: {
 		flex: 1,
 		paddingVertical: 10,
@@ -107,36 +100,11 @@ const styles = StyleSheet.create({
 		padding: 10,
 		marginBottom: 5
 	},
-	detailRowInfoItem: {
-		flex: 1,
-		flexDirection: 'row',
-		borderBottomWidth: StyleSheet.hairlineWidth,
-		borderColor: colors.borderColor,
-		marginBottom: 10,
-		paddingBottom: 5
-	},
-	noBorderBottom: {
-		borderBottomWidth: 0
-	},
 	detailRowText: {
 		flex: 1,
 		fontSize: 12,
 		color: colors.fontSecondary,
 		...fontStyles.normal
-	},
-	alignLeft: {
-		textAlign: 'left'
-	},
-	alignRight: {
-		textAlign: 'right'
-	},
-	viewOnEtherscan: {
-		fontSize: 14,
-		color: colors.primary,
-		...fontStyles.normal,
-		textAlign: 'center',
-		marginTop: 15,
-		marginBottom: 10
 	},
 	hash: {
 		fontSize: 12
@@ -269,71 +237,6 @@ class TransactionElement extends PureComponent {
 		}
 	};
 
-	renderTxDetails = (tx, blockExplorer) => {
-		const {
-			transaction: { gas, gasPrice, value, to, from },
-			transactionHash
-		} = tx;
-		const gasBN = hexToBN(gas);
-		const gasPriceBN = hexToBN(gasPrice);
-		const amount = hexToBN(value);
-		const { conversionRate, currentCurrency } = this.props;
-		const totalGas = isBN(gasBN) && isBN(gasPriceBN) ? gasBN.mul(gasPriceBN) : toBN('0x0');
-		const total = isBN(amount) ? amount.add(totalGas) : totalGas;
-
-		return (
-			<View style={styles.detailRowWrapper}>
-				{this.renderTxHash(transactionHash)}
-				<Text style={styles.detailRowTitle}>{strings('transactions.from')}</Text>
-				<View style={[styles.detailRowInfo, styles.singleRow]}>
-					<Text style={styles.detailRowText}>{renderFullAddress(from)}</Text>
-				</View>
-				<Text style={styles.detailRowTitle}>{strings('transactions.to')}</Text>
-				<View style={[styles.detailRowInfo, styles.singleRow]}>
-					<Text style={styles.detailRowText}>{renderFullAddress(to)}</Text>
-				</View>
-				<Text style={styles.detailRowTitle}>{strings('transactions.details')}</Text>
-				<View style={styles.detailRowInfo}>
-					<View style={styles.detailRowInfoItem}>
-						<Text style={[styles.detailRowText, styles.alignLeft]}>{strings('transactions.amount')}</Text>
-						<Text style={[styles.detailRowText, styles.alignRight]}>
-							{renderFromWei(value)} {strings('unit.eth')}
-						</Text>
-					</View>
-					<View style={styles.detailRowInfoItem}>
-						<Text style={[styles.detailRowText, styles.alignLeft]} />
-						<Text style={[styles.detailRowText, styles.alignRight]}>{hexToBN(gas).toNumber()}</Text>
-					</View>
-					<View style={styles.detailRowInfoItem}>
-						<Text style={[styles.detailRowText, styles.alignLeft]}>
-							{strings('transactions.gas_price')}
-						</Text>
-						<Text style={[styles.detailRowText, styles.alignRight]}>{toGwei(gasPrice)}</Text>
-					</View>
-					<View style={styles.detailRowInfoItem}>
-						<Text style={[styles.detailRowText, styles.alignLeft]}>{strings('transactions.total')}</Text>
-						<Text style={[styles.detailRowText, styles.alignRight]}>
-							{renderFromWei(total)} {strings('unit.eth')}
-						</Text>
-					</View>
-					<View style={[styles.detailRowInfoItem, styles.noBorderBottom]}>
-						<Text style={[styles.detailRowText, styles.alignRight]}>
-							{weiToFiat(total, conversionRate, currentCurrency).toUpperCase()}
-						</Text>
-					</View>
-				</View>
-				{tx.transactionHash &&
-					blockExplorer && (
-						<TouchableOpacity
-							onPress={this.viewOnEtherscan} // eslint-disable-line react/jsx-no-bind
-						>
-							<Text style={styles.viewOnEtherscan}>{strings('transactions.view_on_etherscan')}</Text>
-						</TouchableOpacity>
-					)}
-			</View>
-		);
-	};
-
 	render = () => {
 		const { tx, selected, selectedAddress, conversionRate, currentCurrency, blockExplorer } = this.props;
 		const incoming = toChecksumAddress(tx.transaction.to) === selectedAddress;
@@ -378,7 +281,7 @@ class TransactionElement extends PureComponent {
 						</View>
 					</View>
 				</View>
-				{selected ? this.renderTxDetails(tx, blockExplorer) : null}
+				{selected ? <TransactionDetails transactionObject={tx} blockExplorer={blockExplorer} /> : null}
 			</TouchableOpacity>
 		);
 	};
